tests: Remove test for retaining unknown bits in termios flags
